MAINT: refactor clustal format writer
[NEW] added optional argument `order` to `clustal_from_alignment` to allow
     reordering of sequences in the output.

[CHANGED] clustal_from_alignment now takes a dictionary of sequences
     instead of an alignment object. This matches the other sequence
     format writers.

[CHANGED] renamed `cogent3.format.alignment` to `cogent3.format.sequence`
     to better reflect its purpose.
