+++ changelog.d/20250406_125550_Gavin.Huttley.md
@@ -53,3 +53,5 @@ Uncomment the section that is right (remove the HTML comment wrapper).
   - `subset()`
   - `update()`
   - `union()`
+- `cogent3.format.alignment` has been renamed `cogent3.format.sequence`.
+   This is a backwards incompatible change!
