+++ src/cogent3/format/sequence.py
